image type upload and remove

+++ Modules/Product/Services/ProductImageRepository.php
@@ -2,7 +2,6 @@
 
 namespace Modules\Product\Services;
 
-use Illuminate\Support\Facades\DB;
 use Modules\Core\Traits\FileManager;
 use Modules\Product\Entities\Product;
 use Modules\Product\Entities\ProductImage;
@@ -10,16 +9,13 @@ use Modules\Product\Entities\ProductImage;
 class ProductImageRepository
 {
     use FileManager;
-    protected  $folder_path;
+    protected  $folder_path,$product_image_path;
     private $folder = 'product';
-
     public function __construct()
     {
-        $this->folder_path = storage_path('app/public/images/') . $this->folder . DIRECTORY_SEPARATOR;
+        $this->folder_path = storage_path('app/public/');
+        $this->product_image_path = 'images/product/';
     }
-
-
-
     public function uploadProductImages($product)
     {
         try {
@@ -29,21 +25,15 @@ class ProductImageRepository
                 $files = request()->file('images');
                 foreach ($files as $file) {
                     $fileName = $this->getFileName($file);
-                    $path = 'images/'.$this->folder.'/'.$fileName;
-                    $file->move($this->folder_path, $fileName);
+                    $file->move($this->folder_path.$this->product_image_path, $fileName);
                     $image_type_array = $this->getImageType($product);
-
-
                     $productImage = ProductImage::create(array_merge([
                             'product_id' => $product->id,
-                            'path' => $path
+                            'path' => $this->product_image_path.$fileName
                     ],$image_type_array));
                     $productImageIds[] = $productImage->id;
-
                 }
-
             }
-
             return $productImageIds;
         } catch (\Exception $exception) {
             throw  $exception;
@@ -56,8 +46,8 @@ class ProductImageRepository
         $productImages = $product->images;
         if(isset($productImages) && $productImages->count()>0){
             foreach ($productImages as $productImage){
-                if (file_exists($this->folder_path . $productImage->image)) {
-                    unlink(($this->folder_path . $productImage->image));
+                if (file_exists($this->folder_path . $productImage->path)) {
+                    unlink(($this->folder_path . $productImage->path));
                 }
                 $productImage->delete();
             }
@@ -69,9 +59,9 @@ class ProductImageRepository
 
     public function removeParticularProductImage($productImage):bool
     {
-        if ($productImage) {
-            if (file_exists($this->folder_path . $productImage->image)) {
-                unlink(($this->folder_path . $productImage->image));
+        if ($productImage && isset($productImage->path) && $productImage->path != '') {
+            if (file_exists($this->folder_path . $productImage->path)) {
+                unlink(($this->folder_path . $productImage->path));
             $productImage->delete();
             return true;
             }
@@ -82,7 +72,6 @@ class ProductImageRepository
     private function getImageType(Product $product):array
     {
         $image_type_array= [];
-        $firstImageExist =  true;
         $productImagesCount = ProductImage::where('product_id',$product->id)->count();
         if($productImagesCount == 0) {
             return $image_type_array = [
